+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Feature) (Platform) Generate GRPC Gateway Code
 - (Feature) (Platform) Identity Endpoint
 - (Feature) (Platform) Authz V1 Types
+- (Maintenance) Allow GRPC Marshal Opts
 
 ## [1.2.46](https://github.com/arangodb/kube-arangodb/tree/1.2.46) (2025-02-24)
 - (Bugfix) Clean Phase change properly during upgrade